feat(web): lift graph ordering into engine; GET /graph tree
Move flatten_forest, compare_siblings and the GraphNode/GraphSort/SortKey
types from tui into engine::graph so TUI and web share one ordering
implementation (convention principle 6). The TUI re-points at the moved
code with no change to rendered ordering; the graph ordering tests move
with the functions. Add GET /graph rendering the ordered nodes as a
topologically-sorted nested <ul> tree under GraphSort::default(), with
diamonds drawn once and cycles terminating per the existing logic.

Also corrects the iteration body, which carried a copy-pasted OAuth
description from ITERATION-238.

Implements ITERATION-237 / STORY-179.
