+++ docs/stories/STORY-179-lift-flatten-forest-compare-siblings-into-engine-get-graph-tree-render.md
@@ -1,7 +1,7 @@
 ---
 title: Lift flatten_forest/compare_siblings into engine; GET /graph tree render
 type: story
-status: accepted
+status: complete
 author: jkaloger
 date: 2026-06-30
 tags: []
